Module mathvis::evaluation_tree

This module defines Evaluatable trait as well as types implementing it that are intended to be used.

EvaluationTree is a go-to type to use during your evaluations. Strictly-speaking, it can be anything, and might even not contain actual Nodes (For now it does contains a single root node though).

EvaluationTree can be created by parsing with EvaluationTree::from_tokens function. If you need some custom structure, EvaluationTree can be ignored and bare Nodes used instead, as they both implement Evaluatable anyway.

Modules

args

This module defines Args object and all of the operations that can be performed with it.

functions

This module defines function-related objects that can be used as Node.

Structs

Constant

Basic type implementing Evaluatable. Requires no variables to evaluate and always returns the same value.

EvaluationTree

Represents parsed expression, ready to be evaluated.

Variable

Basic type implementing Evaluatable, having a name. Requires a single variable argument to be evaluated.

Traits

Evaluatable

A trait, defining common interface for everything that can be evaluated.

Operator

A trait defining common evaluatable behavior. Simplifies Evaluatable implementation for types that contain inner node and require no additional parameters.

Type Aliases

Node

Commonly used type through a crate. Represents “tree node” or th.

NodeRef

Commonly used type through a crate. Represents “tre node reference” or th