Module mathvis::evaluation_tree::args

This module defines Args object and all of the operations that can be performed with it.

Args represents the way identifiers should be resolved during evaluation. Currently, there are three valid uses for identifiers (corresponding with variants of ArgType):

variable

function (single argument)

function2 (two arguments)

To actually evaluate the tree, runtime must know values that variables are actually equal to and exact operations represented by functions.

Variable/function can be in tree states regarding to some Args object:

unregistered: Args doesn’t know it. In case it’s required at evaluation, MissingError::MissingEntry error will be returned.

registered: Args knows it exists, but doesn’t have a definition for it yet. Used by trees to communicate variables they need. In case it’s required at evaluation, MissingError::MissingValue error will be returned.

assigned: Args knows about it and has a definition. This is the only state of variable/function that won’t cause an error if asked for at evaluation.

Variables/functions can be assigned to multiple times, this would overwrite them each time.

Variables/functions can be unregistered too. Although, there’s not much use for it apart from debugging.

Structs

Args

Represents argument list as well as their values during evaluation with crate::evaluation_tree::Evaluatable::evaluate function. This object is required to perform it.

ArgsErased

An erased form of Args that does not Display or Eq by actual argument values/states.

ArgsRefErased

An erased reference to Args that does not Display or Eq by actual argument values/states.

Enums

ArgType

Represents types of arguments.

MissingError

Represents error that had happened during evaluation.