Struct mathvis::evaluation_tree::args::Args

pub struct Args { /* private fields */ }

Represents argument list as well as their values during evaluation with crate::evaluation_tree::Evaluatable::evaluate function. This object is required to perform it.

Implementations

impl Args

pub fn new() -> Self

Creates empty Args object.

pub fn erased(self) -> ArgsErased

Converts object into erased form.

pub fn erased_ref(&self) -> ArgsRefErased<'_>

Gets erased reference to an object.

pub fn register_variable(&mut self, name: impl Into<String>)

Registers (but not assigns!) variable with a certain name.

pub fn register_function(&mut self, name: impl Into<String>)

Registers (but not assigns!) function with a certain name.

pub fn register_function2(&mut self, name: impl Into<String>)

Registers (but not assigns!) function2 with a certain name.

pub fn unregister_variable(&mut self, name: impl AsRef<str>)

Unregisters variable with a certain name, if present.

pub fn unregister_function(&mut self, name: impl AsRef<str>)

Unregisters function with a certain name, if present.

pub fn unregister_function2(&mut self, name: impl AsRef<str>)

Unregisters function2 with a certain name, if present.

pub fn assign_variable<N: AsRef<str> + Into<String>>( &mut self, name: N, value: Complex64 )

Assigns to a variable with certain name (and registers it, if necessary)

pub fn assign_function<N: AsRef<str> + Into<String>>( &mut self, name: N, value: impl Fn(Complex64) -> Complex64 + 'static )

Assigns to a function with certain name (and registers it, if necessary)

pub fn assign_function2<N: AsRef<str> + Into<String>>( &mut self, name: N, value: impl Fn(Complex64, Complex64) -> Complex64 + 'static )

Assigns to a function2 with certain name (and registers it, if necessary)

pub fn get_variable<N: AsRef<str> + Into<String>>( &self, name: N ) -> Result<&Complex64, MissingError>

Retrieves reference to a value of variable. May fail with MissingError, if variable is not registered or not assigned to.

pub fn get_function<N: AsRef<str> + Into<String>>( &self, name: N ) -> Result<&dyn Fn(Complex64) -> Complex64, MissingError>

Retrieves reference to a function. May fail with MissingError, if function was not defined.

pub fn get_function2<N: AsRef<str> + Into<String>>( &self, name: N ) -> Result<&dyn Fn(Complex64, Complex64) -> Complex64, MissingError>

Retrieves reference to a function2. May fail with MissingError, if function2 was not defined.

pub fn merge(&mut self, other: Self)

! THIS OPERATION IS NOT SYMMETRIC

Merges this Args object with another. Should be noted, that only registered variable sets merge, but states do not - this always prefers state of a variable in this instance.

pub fn variables(&self) -> impl IntoIterator<Item = &String>

Gets iterator over registered variable names.

pub fn variables_mut( &mut self ) -> impl IntoIterator<Item = (&String, &mut Option<Complex64>)>

Gets iterator over registered variable names as well as their values - Option::None represents no value, while Option::Some represents assigned variable.

pub fn functions(&self) -> impl IntoIterator<Item = &String>

Gets iterator over registered function names.

pub fn functions_mut( &mut self ) -> impl IntoIterator<Item = (&String, &mut Option<Box<dyn Fn(Complex64) -> Complex64>>)>

Gets iterator over registered function names as well as their definitions - Option::None represents no definition, while Option::Some represents defined function.

pub fn functions2(&self) -> impl IntoIterator<Item = &String>

Gets iterator over registered function2 names.

pub fn functions2_mut( &mut self ) -> impl IntoIterator<Item = (&String, &mut Option<Box<dyn Fn(Complex64, Complex64) -> Complex64>>)>

Gets iterator over registered function2 names as well as their definitions - Option::None represents no definition, while Option::Some represents defined function2.