Trait mathvis::evaluation_tree::Evaluatable

pub trait Evaluatable {
    type Res;

    // Required methods
    fn evaluate(&self, arguments: &Args) -> Result<Self::Res, MissingError>;
    fn args(&self) -> Args;
}

A trait, defining common interface for everything that can be evaluated.

Required Associated Types

type Res

Type of the evaluation result.

Required Methods

fn evaluate(&self, arguments: &Args) -> Result<Self::Res, MissingError>

Attempts to evaluate this object, possibly failing with MissingError to indicate missing argument.

fn args(&self) -> Args

Retrieves arguments this object requires to be evaluated.